+++ packages/pocket-ic/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New functions `auto_progress` and `stop_progress` to make IC instances
   progress (updating time and executing rounds) automatically.
 - New subnet specification allowing to set very high instruction limits for (asymptotic) benchmarking canister code.
+- New field `dts_flag` in `SubnetSpec` controlling if DTS is enabled (enabled by default on all non-benchmarking subnets).
 
 ## Changed
 - `get` and `post` helpers which are used by all server-facing functions now poll on results, because 1) instances can be busy with other computations and 2) the `post`ed computations may take longer than the specified timeout or the `reqwest` client's own timeout. With this change, very long-running computations can be handled by the library.